move ShellUtils and NetworkUtils to utils dir
