+++ heron/schedulers/src/java/com/twitter/heron/scheduler/aurora/AuroraController.java
@@ -20,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
 
-import com.twitter.heron.spi.common.ShellUtils;
+import com.twitter.heron.spi.utils.ShellUtils;
 
 /**
  * This file defines Utils methods used by Aurora
