+++ heron/spi/tests/java/com/twitter/heron/spi/utils/SchedulerUtilsTest.java
@@ -24,7 +24,6 @@ import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.twitter.heron.common.basics.FileUtils;
 import com.twitter.heron.proto.system.Common;
-import com.twitter.heron.spi.common.ShellUtils;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({FileUtils.class, ShellUtils.class, SchedulerUtils.class})
